+++ src/main/java/lagatrix/server/entities/dto/partition/Partition.java
@@ -1,4 +1,4 @@
-package lagatrix.server.entities.devices;
+package lagatrix.server.entities.dto.partition;
 
 import java.util.Objects;